Create Cave.py
I have used Turtle graphics in this projects. In the main loop i have avoided the use of sleep function for delay between each movement of walls as to let the main pointer move more freely. Else it will move with a lag. The maximum memory usage is observed to be upto 24mb.